requirements.txt
alembic==1.16.5
async-timeout==5.0.1
blinker==1.9.0
click==8.3.0
colorama==0.4.6
Flask==3.1.2
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
psycopg2-binary==2.9.10
python-dotenv==1.1.1
redis==6.4.0
SQLAlchemy==2.0.43
tomli==2.2.1
typing_extensions==4.15.0
Werkzeug==3.1.3
alembic==1.16.5
async-timeout==5.0.1
blinker==1.9.0
click==8.3.0
colorama==0.4.6
Flask==3.1.2
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
psycopg2==2.9.11
psycopg2-binary==2.9.10
python-dotenv==1.1.1
redis==6.4.0
SQLAlchemy==2.0.43
tomli==2.2.1
typing_extensions==4.15.0
Werkzeug==3.1.3
